refactor: :truck: firebase関連のコードをsharedディレクトリに移動
