+++ packages/app/src/components/buttons/SinglePlayButton.tsx
@@ -1,6 +1,6 @@
 import { useScene } from "@/game/hooks/useScene";
 import { useUpdatePlayers } from "@/game/hooks/useUpdatePlayers";
-import { analytics } from "@/lib/firebase/config";
+import { analytics } from "@/shared/firebase/config";
 import { logEvent } from "firebase/analytics";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
