+++ packages/app/src/shared/firebase/config.ts
